+++ src/components/bookmark/BookmarkedServices.jsx
@@ -17,7 +17,7 @@ export const BookmarkedServices = () => {
   const { user: currentUser } = useAuth();
   const { data: services, isLoading } = useQuery({
     queryKey: ["services"],
-    queryFn: getServices,
+    queryFn: () => getServices,
     enabled: !!currentUser,
   });
 
@@ -62,7 +62,6 @@ const BookmarkedServicesCard = ({
   setCachedServices,
   cachedServices,
 }) => {
-
   const { user: currentUser } = useAuth();
 
   const hasBookmarkedService = service?.likes?.find(